+++ tests/spintwo/test_spintwo_wavefunction.py
@@ -285,7 +285,7 @@ def test_phase_all():
         np.ones((64, 64), dtype="complex128"),
     )
 
-    phase = np.random.uniform(size=(64, 64), dtype=np.float64)
+    phase = np.random.uniform(size=(64, 64))
     wavefunction.apply_phase(phase, "all")
 
     np.testing.assert_allclose(np.angle(wavefunction.plus2_component), phase)
@@ -306,7 +306,7 @@ def test_phase_multiple_components():
         np.ones((64, 64), dtype="complex128"),
     )
 
-    phase = np.random.uniform(size=(64, 64), dtype=np.float64)
+    phase = np.random.uniform(size=(64, 64))
     wavefunction.apply_phase(phase, ["plus1", "minus1"])
 
     np.testing.assert_allclose(np.angle(wavefunction.plus1_component), phase)
@@ -324,7 +324,7 @@ def test_phase_single():
         np.ones((64, 64), dtype="complex128"),
     )
 
-    phase = np.random.uniform(size=(64, 64), dtype=np.float64)
+    phase = np.random.uniform(size=(64, 64))
     wavefunction.apply_phase(phase, "zero")
 
     np.testing.assert_allclose(np.angle(wavefunction.zero_component), phase)
